+++ src/backend/api/Fusion.Resources.Api/Controllers/Person/PersonController.cs
@@ -154,15 +154,6 @@ namespace Fusion.Resources.Api.Controllers
             return respObject;
         }
 
-        [HttpGet("/persons/resource-owners")]
-        public async Task<ActionResult<List<ApiDepartment>>> GetResourceOwners([FromQuery(Name = "q")] string query)
-        {
-            var request = new SearchResourceOwners(query);
-            var result = await DispatchAsync(request);
-
-            return Ok(result.Select(x => new ApiDepartment(x)));
-        }
-
         private async Task<string?> ResolveSector(string department)
         {
             var request = new GetDepartmentSector(department);
