+++ src/backend/api/Fusion.Resources.Domain/Models/LineOrg/PaginatedResponse.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Text.Json.Serialization;
 
 namespace Fusion.Resources.Domain.LineOrg
 {
@@ -6,6 +7,9 @@ namespace Fusion.Resources.Domain.LineOrg
     {
         public int Count { get; set; }
         public int TotalCount { get; set; }
-        public List<T> Value { get; set; }
+
+        [JsonPropertyName("@NextPage")]
+        public string? NextPage { get; set; }
+        public List<T> Value { get; set; } = new List<T>();
     }
 }
